Cache RBS Environment object
To optimize the loading RBS Environment, this caches the resolved RBS
environment object.  It will improve the performance of loading it
twice if the cache is available (3.925s to 1.865s).  The cache will be
invalidated automatically if the type definitions are changed.

As a penalty, the loading performance will be decreased twice if the
cache is unavailable (3.925s to 7.981s) because generating cache file
via `RBS::Writer` is very slow.